+++ components/Hero/HeroSignIn/HeroSignIn.tsx
@@ -1,4 +1,4 @@
-import TextInput from "@/components/FormControllers/TextInput";
+import TextInput from "@/components/placeholderfc/TextInput";
 import SnackbarAlert from "@/components/SnackbarAlert/SnackbarAlert";
 import { SNACKBAR_ALERT_INITIAL } from "@/helpers/forms";
 import { PAGES } from "@/helpers/navigation";
